Added QueueExercises from https://stepik.org/lesson/77929/step/1?unit=54781 created maven project and tests
